Spellchecker webservice init commit
