+++ go.mod
@@ -6,12 +6,14 @@ require (
 	github.com/RoaringBitmap/roaring v0.4.20
 	github.com/alldroll/cdb v1.0.2
 	github.com/alldroll/go-datastructures v0.0.0-20190322060030-1d3a19ff3b29
+	github.com/djimenez/iconv-go v0.0.0-20160305225143-8960e66bd3da
 	github.com/edsrzf/mmap-go v0.0.0-20190108065903-904c4ced31cd
 	github.com/gorilla/handlers v1.4.0
 	github.com/gorilla/mux v1.7.1
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.3 // indirect
+	golang.org/x/net v0.0.0-20191119073136-fc4aabc6c914
 	golang.org/x/sync v0.0.0-20190423024810-112230192c58
 	golang.org/x/sys v0.0.0-20190426135247-a129542de9ae // indirect
 )
