+++ pkg/cmd/spellchecker/cmd/root.go
@@ -1,0 +1,32 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var (
+	configPath string
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "spellchecker",
+	Short: "cli to interact with a spellchecker service",
+	Long:  `cli to interact with a spellchecker service`,
+}
+
+func init() {
+	rootCmd.PersistentFlags().StringVarP(&configPath, "config", "c", "", "path to the config file")
+	rootCmd.MarkPersistentFlagFilename("config")
+	rootCmd.MarkPersistentFlagRequired("config")
+}
+
+// Execute runs commands handling
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
