+++ pkg/cmd/spellchecker/cmd/spellchecker.go
@@ -1,0 +1,1 @@
+package cmd
